docs #13 - next.config 및 tailwind.config 파일 구조 변경 및 확장자명 변경
