--- next.config.js
@@ -1,13 +1,0 @@
-const nextConfig = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'http',
-        hostname: '**'
-      },
-      { protocol: 'https', hostname: '**' }
-    ]
-  }
-};
-
-module.exports = nextConfig;
