+++ next.config.mjs
@@ -1,0 +1,15 @@
+import withPlaiceholder from '@plaiceholder/next';
+
+const nextConfig = {
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'http',
+        hostname: '**'
+      },
+      { protocol: 'https', hostname: '**' }
+    ]
+  }
+};
+
+export default withPlaiceholder(nextConfig);
